MASSIVE performance boost.
So here's the deal. Every single time dc-tool saw that no packet was
incoming from the DC, it would sleep FOR A WHOLE GODDAMN HALF SECOND,
which absolutely DESTROYED performance and would make it less stable
too, as the chances of packet loss were higher.

This change removes any sleeping by default (we should be processing
them as quickly as they come), BUT will do a sleep of 0ns if the
SAVE_MY_FANS flag is enabled--whichi is simply going to be equivalent to
yielding the thread so we don't got all the CPU.